Moving interfaces to the implementing file
