+++ upcloud/service/object_storage.go
@@ -8,6 +8,16 @@ import (
 	"github.com/UpCloudLtd/upcloud-go-api/upcloud/request"
 )
 
+type ObjectStorage interface {
+	GetObjectStorages() (*upcloud.ObjectStorages, error)
+	GetObjectStorageDetails(r *request.GetObjectStorageDetailsRequest) (*upcloud.ObjectStorageDetails, error)
+	CreateObjectStorage(r *request.CreateObjectStorageRequest) (*upcloud.ObjectStorageDetails, error)
+	ModifyObjectStorage(r *request.ModifyObjectStorageRequest) (*upcloud.ObjectStorageDetails, error)
+	DeleteObjectStorage(r *request.DeleteObjectStorageRequest) error
+}
+
+var _ ObjectStorage = (*Service)(nil)
+
 // GetObjectStorages returns the available objects storages
 func (s *Service) GetObjectStorages() (*upcloud.ObjectStorages, error) {
 	objectStorages := upcloud.ObjectStorages{}
